requirements.txt
asgiref==3.5.2
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
breezy==3.0.2
cachetools==5.2.0
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloud-init==22.3.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
docutils==0.16
dulwich==0.19.15
empy==3.3.2
entrypoints==0.3
fastimport==0.9.8
filelock==3.8.0
Flask==2.2.2
fonttools==4.38.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.57.0
gpg===1.13.1-unknown
gunicorn==20.1.0
httplib2==0.21.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==5.0.0
importlib-resources==5.12.0
incremental==16.10.1
ipython==7.13.0
ipython_genutils==0.2.0
itsdangerous==2.1.2
jedi==0.15.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.1
matplotlib==3.7.0
more-itertools==4.2.0
mpi4py==3.0.3
mysqlclient==2.1.1
netifaces==0.10.4
nose==1.3.7
numpy==1.24.2
oauthlib==3.1.0
olefile==0.46
packaging==23.0
pandas==1.5.3
paramiko==2.6.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
plotly==5.13.1
prompt-toolkit==2.0.10
protobuf==4.21.12
psutil==5.5.1
psycopg2==2.9.5
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-snappy==0.5.3
PyTrie==0.2
pytz==2022.7.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
rosdep==0.22.1
rosdep-modules==0.22.1
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosinstall==0.7.8
rosinstall-generator==0.1.22
rospkg==1.4.0
rospkg-modules==1.4.0
rsa==4.9
seaborn==0.12.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sos==4.4
sqlparse==0.4.3
ssh-import-id==5.10
systemd-python==234
tenacity==8.2.1
traitlets==4.3.3
Twisted==18.9.0
txaio==2.10.0
u-msgpack-python==2.1
ubuntu-advantage-tools==27.11.3
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
vcstools==0.1.42
virtualenv==20.16.5
wadllib==1.3.3
wcwidth==0.1.8
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
zipp==3.15.0
zope.interface==4.7.1
os.path